Sat, May 18, 2024, 8:50 PM -03:00

+++ week3/problems/runoff/runoff.c
@@ -184,7 +184,7 @@ bool print_winner(void)
     }
 
     // if there is one winner with more than half of the votes
-    if (votes > (voter_count / 2))
+    if (votes > ((float) (voter_count / 2)))
     {
         printf("%s wins!\n", contender);
         return true;
